Optimize TDBTrace >> #traceAtTime:ifAbsent:
- Early exit if timeIndex it out of the current trace's range
- Exploit the the OrderedDictionary's internal structure for greater performance

In a larger tracing exercise for #15, this method still takes 8% of the time (as opposed to 39%) before.

Note that we have still the possibility to implement an inline binary search; however; a large majority of all traces has <= 4 children only (measured 95% - 98.5% for different domains, avg length ~ 1.00), so this is left as an exercise to the reader.