Implement trace-spawning, which opens another window on the same trace in another window
The trace and the memory are not (yet?) copied, just the cursor, and the process remains the same. Manage a simple reference list of views on the process to terminate it exactly once when closing the last debugger. Relates #29.

Fixes handling of bottomContexts when updating the context stack/the selected path.