memory simulator: fix primitiveClone
In the past, we accidentally would have re-executed primitive 148 to access the current value for each slot. The primitive would not fail but copy the requested instead. Harden test data for the range retracing simulator to avoid similar incidents in the future (this test, however, does not trigger the defect code path because it uses primitiveRetracingTimeIndex).