history explorer: add details mode
Choose the representation mode "details" from the upper right corner of the tool and select any context to view the corresponding value. The value can be a Text or a Form or otherwise is displayed as a print-string. If the context has any expanded children, all values from the children will be represented. If the value frays out during displaying, all frayed out results will be presented.

Further improvements include:
- compaction option to shrink columns in list mode
- show wait cursor during computation
- fix text-spawn (cmd + o)
- handle errors/frayouts in window title
- overall refactoring of labels and error handling
- configurable styles for error/header via uitheme
- add many acceptance tests for history explorer

Other changes:
- move up TraceDebugger>>#attributesForStyle: to Movel>>#tdbAttributesForStyle:
- add Object>>#isTdbIsolatedError
- add Cursor>>#debug:atTime: